clarify: drop redundant casts on mXrefSize, document safe pagesCount cast
mXrefSize is already ObjectIDType (unsigned long), so the explicit
(unsigned long) and (unsigned long long) casts on it added nothing.
Add a comment on the mPagesCount cast line noting that the bounds
check above guarantees the cast cannot truncate.

+++ PDFWriter/PDFParser.cpp
@@ -884,13 +884,14 @@ EStatusCode PDFParser::ParsePagesObjectIDs()
 		// reject negative counts and counts larger than the xref table can possibly hold.
 		// every leaf page is an indirect object that must have an xref entry, so the
 		// declared /Count cannot legitimately exceed the xref size.
-		if(pagesCountValue < 0 || (unsigned long long)pagesCountValue > (unsigned long long)mXrefSize)
+		if(pagesCountValue < 0 || (unsigned long long)pagesCountValue > mXrefSize)
 		{
-			TRACE_LOG2("PDFParser::ParsePagesObjectIDs, invalid pages count %lld (xref size %lu)", pagesCountValue, (unsigned long)mXrefSize);
+			TRACE_LOG2("PDFParser::ParsePagesObjectIDs, invalid pages count %lld (xref size %lu)", pagesCountValue, mXrefSize);
 			status = PDFHummus::eFailure;
 			break;
 		}
 
+		// safe: bounded above by mXrefSize (an unsigned long) by the check above
 		mPagesCount = (unsigned long)pagesCountValue;
 		mPagesObjectIDs = new ObjectIDType[mPagesCount];
